Change shortcut link to AutoJIT Homescreen Edition
The instructions for the shortcut are in the discord, but it should be pretty simple to figure it out. I can add the instructions to a text element inside of the shortcut before merging this change tonight if that's helpful. All you have to do is press select app, choose you app, and add the shortcut to homescreen.